pyproject.toml
[tool.poetry]
name = "HTTP Conformance"
version = "0.1.0"
description = "HTTP Conformance Test Suite"
authors = ["Jannis Rautenstrauch <jannis.rautenstrauch@cispa.de>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "1.5.3"
Jinja2 = "3.1.2"
astor = "0.8.1"
urllib3 = "1.26.12"
validators = "0.20.0"
Flask = "2.1.3"
watchdog = "2.1.9"
mitmproxy = "8.1.1"
StrEnum = "0.4.8"
requests = "^2.28.1"
tranco = "0.6"
tld = "0.12.6"
peewee = "3.15.3"
httpx = {extras = ["http2", "cli"], version = "0.23.0"}
dpkt = "1.9.8"
httptools = "0.5.0"
scapy = "2.4.5"
psycopg2 = "2.9.5"
tqdm = "4.64.1"
python-dotenv = "0.21.0"
matplotlib = "3.6.3"
jupyterlab = "3.5.3"
lxml = "4.9.2"
scikit-learn = "1.3.2"
werkzeug = "2.2.2"

[tool.poetry.dev-dependencies]
ipykernel = "^6.16.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
